Create mountdata function
In order to be able to synchronize data mount and Dropbox
process, this function was created.

When data is mounted, the Dropbox process can be started - that's
because the dropbox dir is inside the data mount point. Actually, a
symbolic link is created in ~/Dropbox to point to ~/data/Dropbox.

Hence, the premise that states every data file will not reside in
the local disk (and only in the remote disk) is achieved.

This solution is for Dropbox, but it can be extended for each
cloud services that maps directories in the local disk - those
directories will be mapped to the external disk, when mounted.